// src/analysis/types.ts
export interface AnalysisLimits {
  maxInventoryFiles: number;
  maxAnalyzedFiles: number;
  maxScannedBytes: number;
  maxSymbols: number;
  maxRelationships: number;
}

export const DEFAULT_ANALYSIS_LIMITS: AnalysisLimits = {
  maxInventoryFiles: 20_000,
  maxAnalyzedFiles: 5_000,
  maxScannedBytes: 64 * 1024 * 1024,
  maxSymbols: 100_000,
  maxRelationships: 250_000
};